+++ FBSnapshotsViewer/Test Results/Interactor/TestResultsInteractorIO.swift
@@ -13,9 +13,11 @@ protocol TestResultsInteractorInput: class, AutoMockable {
     func openInKaleidoscope(testResult: SnapshotTestResult)
     func openInXcode(testResult: SnapshotTestResult)
     func accept(testResult: SnapshotTestResult)
+    func reject(testResult: SnapshotTestResult)
     func copy(testResult: SnapshotTestResult)
 }
 
 protocol TestResultsInteractorOutput: class, AutoMockable {
     func didFailToAccept(testResult: SnapshotTestResult, with error: Error)
+    func didFailToReject(testResult: SnapshotTestResult, with error: Error)
 }