+++ FBSnapshotsViewer/Test Results/User Interface/TestResultCell.swift
@@ -13,6 +13,7 @@ protocol TestResultCellDelegate: class, AutoMockable {
     func testResultCell(_ cell: TestResultCell, viewInXcodeButtonClicked: NSButton)
     func testResultCell(_ cell: TestResultCell, viewInKaleidoscopeButtonClicked: NSButton)
     func testResultCell(_ cell: TestResultCell, acceptSnapshotsButtonClicked: NSButton)
+    func testResultCell(_ cell: TestResultCell, rejectSnapshotButtonClicked: NSButton)
     func testResultCell(_ cell: TestResultCell, copySnapshotButtonClicked: NSButton)
 }
 
@@ -25,6 +26,7 @@ class TestResultCell: NSCollectionViewItem {
     @IBOutlet private weak var viewInKaleidoscopeButton: NSButton!
     @IBOutlet private weak var viewInXcodeButton: NSButton!
     @IBOutlet private weak var acceptSnapshotsButton: NSButton!
+    @IBOutlet private weak var rejectSnapshotsButton: NSButton!
     @IBOutlet private weak var copySnapshotButton: NSButton!
     
     @IBOutlet private weak var imagesContainerView: NSView!
@@ -139,5 +141,9 @@ class TestResultCell: NSCollectionViewItem {
     
     @objc @IBAction func acceptSnapshotsButtonClicked(_ sender: NSButton) {
         delegate?.testResultCell(self, acceptSnapshotsButtonClicked: sender)
+    }
+
+    @objc @IBAction func rejectSnapshotButtonClicked(_ sender: NSButton) {
+        delegate?.testResultCell(self, rejectSnapshotButtonClicked: sender)
     }
 }